Remove unwrap and panic after rebase. NOT FUNCTIONAL
